Committing because storm.
Update is on its way to completion.  Entering Codes should be sanitized, but not right now.

Delete is working swimmingly, reasonably certain.

I have made the decision to forgo adding Vehicle Class and Files to the Project until all
options are working with the Person class exclusively.  I expect integration to be
straightforward.  Famous Last Words...